Check for textParas that are on the same line when writing out extracted text.
