+++ extractor/text_bound.go
@@ -13,23 +13,10 @@
 package extractor
 
 import (
+	"math"
+
 	"github.com/unidoc/unipdf/v3/model"
 )
-
-var serial serialState
-
-type serialState struct {
-	mark   int
-	word   int
-	strata int
-	line   int
-	para   int
-}
-
-func (serial *serialState) reset() {
-	var empty serialState
-	*serial = empty
-}
 
 /*
  * Sorting functions.
@@ -162,18 +149,40 @@ func overlappedYRect(r0, r1 model.PdfRectangle) bool {
 	return (r0.Lly <= r1.Lly && r1.Lly <= r0.Ury) || (r0.Lly <= r1.Ury && r1.Ury <= r0.Ury)
 }
 
-// minInt return the lesser of `a` and `b`.
-func minInt(a, b int) int {
-	if a < b {
-		return a
+// rectUnion returns the smallest axis-aligned rectangle that contains `b1` and `b2`.
+func rectUnion(b1, b2 model.PdfRectangle) model.PdfRectangle {
+	return model.PdfRectangle{
+		Llx: math.Min(b1.Llx, b2.Llx),
+		Lly: math.Min(b1.Lly, b2.Lly),
+		Urx: math.Max(b1.Urx, b2.Urx),
+		Ury: math.Max(b1.Ury, b2.Ury),
 	}
-	return b
 }
 
-// maxInt return the greater of `a` and `b`.
-func maxInt(a, b int) int {
-	if a > b {
-		return a
+// rectIntersection returns the largest axis-aligned rectangle that is contained by `b1` and `b2`.
+func rectIntersection(b1, b2 model.PdfRectangle) (model.PdfRectangle, bool) {
+	if !intersects(b1, b2) {
+		return model.PdfRectangle{}, false
 	}
-	return b
+	return model.PdfRectangle{
+		Llx: math.Max(b1.Llx, b2.Llx),
+		Urx: math.Min(b1.Urx, b2.Urx),
+		Lly: math.Max(b1.Lly, b2.Lly),
+		Ury: math.Min(b1.Ury, b2.Ury),
+	}, true
+}
+
+// intersects returns true if `r0` and `r1` overlap in the x and y axes.
+func intersects(b1, b2 model.PdfRectangle) bool {
+	return intersectsX(b1, b2) && intersectsY(b1, b2)
+}
+
+// intersectsX returns true if `r0` and `r1` overlap in the x axis.
+func intersectsX(b1, b2 model.PdfRectangle) bool {
+	return b1.Llx <= b2.Urx && b2.Llx <= b1.Urx
+}
+
+// intersectsY returns true if `r0` and `r1` overlap in the y axis.
+func intersectsY(b1, b2 model.PdfRectangle) bool {
+	return b1.Lly <= b2.Ury && b2.Lly <= b1.Ury
 }
